+++ test.js
@@ -1,0 +1,1 @@
+console.log('No tests specified');
